backend/requirements.txt
flask
python-chess
flask-cors
numpy